add Domino.find! method which waits for matching node
When working with asynchronous Capybara drivers, it may be necessary
to wait for dominos to appear eventually. The `find!` method uses
Capybara's `find` instead of `all` to wait for asynchronously created
nodes.

Capybara raises informative error messages if no match is found. This
prevents hard to debug "undefined method of nil" errors when trying to
interact with an expected unique Domino on the page.

* Added Domino.find!
* Added tests for matching, non-matching and missing selector cases
* Added section to README discussing usage with JS enabled Capybara
  drivers