Improve performance of `find_by` and `find_by!`
`find_by` unnecessarily fetches all attributes of a node.  This becomes
especially problematic during certain tests utilizing Capybara, since it
triggers the wait timeout for each missing attribute in the node.

This change alters the logic so that rather then fetching all of the node's
attributes, we only fetch the attributes that we need to test against.